[BUGFIX] Auto Analysis
Bug mentioned in previous commit was in `creait` library. This was due
to early timeout, that was explicitly set during curl handle creation.
Removing the timeout fixes this.

Current commit prints info about renamed binaries in tabular form.

TODO: Renamed binaries are not persistent locally. Need to save rizin
analysis in order to keep the names in server and in rizin analysis in
sync.