[FIX] Adding Cutter Support & MacOS Fix
Moving towards v0.2.x by adding cutter support. The commit performs some
refactoring to reuse already existing code and make it work with rizin.

This commit fixes build error on MacOS. The error was due to direct
access of `RzListIter` members instead of using rizin api for
interacting with `rz_list_iter_t` a.k.a `RzListIter`.

The above issue can be seen fixed in rizin in this commit :
8fba9b41896b9d7ff3c2c4e8ad6808316daf59dc