Create Analysis & Get Analysis Status For Cutter
Cutter is now able to create new analysis by automatically uploading the
currently opened binary file. We can now fetch analysis status as well.
A background thread keeps updating the analysis status in local
database. This thread was already implemented in Rizin plugin.

The analysis bug was not technically bug. The sample taken was wrong, of
an unsupported platform. This means we need to check whether the sample
is of supported platform before creating analysis. This is a refactoring
part however IMO.

Cutter now also loads the already available config, instead of asking
the user to remove the old config to create a new one. This means we can
modify the current config using cutter config setup dialog.

A bug still remains when cutter config is saved, cutter crashes.