REFACTORING : Add `ReaiPluginTable` Wrapper Over `RzTable`
To display auto rename results in both rizin and cutter, this series of
refactor commits will add a new interface named `ReaiPluginTable` that
wraps internal implementation of table in both the plugins. In Rizin,
`RzTable` will be used and in Cutter, a new widget will be implemented.

Currently there's a blocker for this refactor in Rizin. Rizin
maintainers recommended performing requisite changes in the tool and
creating a PR.

This refactor will require resolution of https://github.com/rizinorg/rizin/issues/1961