Hack To Make `ReaiPluginTable` Work.
We can guess the internal API based on what language is being
compiled (C or C++). If it's C then it's Rizin plugin and then we'll use
Rizin's RzTable API, otherwise, we're in Cutter and we need to define an
use a Qt widget to display table.

SEE COMMIT MSG : dc2575eafd177430e41cca6ffaf38c37a613ee3a