`ReaiPluginTable` API Almost Works.
The hack didn't work (switching on #ifdef __cplusplus) to decide the
internal API implementation. Had to manually create a `RzPVector` to
represent rows and insert values. This might not be perfect solution but
it seems to work for now and is now not a blocker like before.

There might be some issues with printing the values using
`reai_plugin_table_add_rowf` but those can be fixed later on. The whole
point was to switch on internal implementation to behave like a TUI at
one place and like a GUI in another.