Working On Batch Rename in Cutter. Batch Rename Bug.
Need to find a one-to-one correspondence between rizin functions and
reveng.ai functions. Using addresses does not seem to work, also
reveng.ai seem to not be using provided function names during creation
of analysis.

UI and other components of function rename seem to be working. If I'm
able to find an injective map, then this is the last part of MVP. Rest
remain bugs only...