[BUGFIX] Cutter Config Save Crash Fixed & Installer
The commit adds the Ubuntu dockerfile, starts work on an installer
script for Unix based distros. This commit also fixes a longstanding bug
on config save. Cutter crashed when saving plugin.

Bug explanation :
The bug was because of freeing of an unallocated (static) memory region.
The bug was present in both the plugin's shared code. The commit also
fixes an unintentionally introduced bug in last commit where when
returning the pointer to rizin plugin's internal state, we didn't check
whether the internal state already exists. This was accidentally removed
in last refactoring.