[BUGFIX] Correct `ReaiPluginTable` Generation
The commit refactors and fixes the table generation bugs in
`ReaiPluginTable` abstraction. The plugin now completely uses the API
provided by `RzTable` to generate the table and display it.

This was possible because of recent merge of PR
https://github.com/rizinorg/rizin/pull/4625

This also means that users might face some problems if they use a
release version of Rizin/Cutter. Users (for now) must build Rizin from
source until unless a new version of Rizin is released with the changes
in PR present.