[HACK] Rizin `RzTable` Temporary Hack To Make `ReaiPluginTable` Work
After the PR to add a new API for RzTable to take a `va_list` and a
format string to add values to RzTable, we cannot use it directly in
here because the code is not released and users won't be able to get the
binary and library files with updated code. For this, I'll introduce a
temporary hack to make the plugin work.

This hack will be removed as soon as I see the next release of Rizin.