Remove Model From Config
Model version and name specified in config is not really a good idea for
future use cases. There can be models for multiple platforms, and
different versions might exist (like they do now) in future.

For this, models are loaded everytime cutter or rizin is opened and when
creating analysis are available through a dropdown in cutter and through
a command "REm" in rizin.