Don't Display Failed Renaming Attempt Results
When applying an existing analysis, or performing an auto analysis
(fetching function name suggestions and renaming), some functions were
failing to be renamed, either because of function not being at specified
address or because of absence of an appropriate match.

This somethimes created very large tables of failed renames, especially
when a binary has lots of functions and only a few were successfully
renamed.

For this, the failed renames are now visible in log only. Frontend
mustn't show failed rename operations in either Rizin or Cutter.