New Analysis Cmd Group, Remove YES/NO Questions & Split To Commands
For better use with rz-pipe, removed yes/no questions from commands.
Such commands are now split into two, where one is for answering yes and
other is for answering no.

In some cases where yes/no was for creating rizin analysis. I've removed
the question and will automatically create a rizin analysis without
asking the user. For the moment, I cannot think of a reason, why user
would resist the plugin creating analysis on their behalf.

The analysis command is now it's own command group. Command for creating
analysis is now `REac` and `REacp` instead of `REa` for creating
analysis in PRIVATE and in PUBLIC respectively.